requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.7
python-dotenv==1.0.0
django-rest-swagger==2.2.0
django-redis==5.3.0

